Ajay-Patil/Similarity_Search/requirements.txt
numpy
pillow
tensorflow
flask